final commit

+++ js/break_and_continue.js
@@ -2,7 +2,7 @@
 
 for (let number = 0; number < 1 || number > 50 || number % 2 === 0 || isNaN(number);) {
 
-    number = parseFloat(prompt(`Enter an odd number between 1 and 50.`))
+    number = parseInt(prompt(`Enter an odd number between 1 and 50.`))
 
     if (number > 0 && number < 50 && number % 2 !== 0) {
